use API infalter as main inflater
