+++ cli_tools/common/image/importer/api_inflater_test.go
@@ -50,18 +50,16 @@ func TestCreateInflater_File(t *testing.T) {
 	facade, ok := inflater.(*inflaterFacade)
 	assert.True(t, ok)
 
-	mainInflater, ok := facade.mainInflater.(*daisyInflater)
 	assert.True(t, ok)
-	assert.Equal(t, "zones/us-west1-c/disks/disk-1234", mainInflater.inflatedDiskURI)
-	assert.Equal(t, "gs://bucket/vmdk", mainInflater.wf.Vars["source_disk_file"].Value)
-	assert.Equal(t, "projects/subnet/subnet", mainInflater.wf.Vars["import_subnet"].Value)
-	assert.Equal(t, "projects/network/network", mainInflater.wf.Vars["import_network"].Value)
+	assert.Equal(t, "zones/us-west1-c/disks/disk-1234", facade.daisyInflater.inflatedDiskURI)
+	assert.Equal(t, "gs://bucket/vmdk", facade.daisyInflater.wf.Vars["source_disk_file"].Value)
+	assert.Equal(t, "projects/subnet/subnet", facade.daisyInflater.wf.Vars["import_subnet"].Value)
+	assert.Equal(t, "projects/network/network", facade.daisyInflater.wf.Vars["import_network"].Value)
 
-	network := getWorkerNetwork(t, mainInflater.wf)
+	network := getWorkerNetwork(t, facade.daisyInflater.wf)
 	assert.Nil(t, network.AccessConfigs, "AccessConfigs must be nil to allow ExternalIP to be allocated.")
 
-	realInflater, _ := facade.shadowInflater.(*apiInflater)
-	assert.NotContains(t, realInflater.guestOsFeatures,
+	assert.NotContains(t, facade.apiInflater.guestOsFeatures,
 		&compute.GuestOsFeature{Type: "UEFI_COMPATIBLE"})
 }
 
@@ -86,12 +84,12 @@ func TestCreateAPIInflater_IncludesUEFIGuestOSFeature(t *testing.T) {
 	request := ImageImportRequest{
 		UefiCompatible: true,
 	}
-	realInflater, _ := createAPIInflater(request, nil, &storage.Client{}, logging.NewToolLogger(t.Name())).(*apiInflater)
-	assert.Contains(t, realInflater.guestOsFeatures,
+	apiInflater := createAPIInflater(request, nil, &storage.Client{}, logging.NewToolLogger(t.Name()), true)
+	assert.Contains(t, apiInflater.guestOsFeatures,
 		&compute.GuestOsFeature{Type: "UEFI_COMPATIBLE"})
 }
 
-func TestAPIInflater_Inflate_CreateDiskFailed_CancelWithoutDeleteDisk(t *testing.T) {
+func TestAPIInflater_ShadowInflate_CreateDiskFailed_CancelWithoutDeleteDisk(t *testing.T) {
 	mockCtrl := gomock.NewController(t)
 	defer mockCtrl.Finish()
 	mockComputeClient := mocks.NewMockClient(mockCtrl)
@@ -101,7 +99,7 @@ func TestAPIInflater_Inflate_CreateDiskFailed_CancelWithoutDeleteDisk(t *testing
 	mockLogger := mocks.NewMockLogger(mockCtrl)
 	mockLogger.EXPECT().Debug("apiInflater.inflate is canceled: cancel")
 
-	inflater := createAPIInflater(ImageImportRequest{
+	apiInflater := createAPIInflater(ImageImportRequest{
 		Source:       fileSource{gcsPath: "gs://bucket/vmdk"},
 		Subnet:       "projects/subnet/subnet",
 		Network:      "projects/network/network",
@@ -109,10 +107,7 @@ func TestAPIInflater_Inflate_CreateDiskFailed_CancelWithoutDeleteDisk(t *testing
 		ExecutionID:  "1234",
 		NoExternalIP: false,
 		WorkflowDir:  daisyWorkflows,
-	}, mockComputeClient, &storage.Client{}, mockLogger)
-
-	apiInflater, ok := inflater.(*apiInflater)
-	assert.True(t, ok)
+	}, mockComputeClient, &storage.Client{}, mockLogger, true)
 
 	// Send a cancel signal in prior to guarantee cancellation logic can be executed.
 	cancelResult := apiInflater.Cancel("cancel")
@@ -122,7 +117,7 @@ func TestAPIInflater_Inflate_CreateDiskFailed_CancelWithoutDeleteDisk(t *testing
 	assert.Error(t, err)
 }
 
-func TestAPIInflater_Inflate_CreateDiskSuccess_CancelWithDeleteDisk(t *testing.T) {
+func TestAPIInflater_ShadowInflate_CreateDiskSuccess_CancelWithDeleteDisk(t *testing.T) {
 	mockCtrl := gomock.NewController(t)
 	defer mockCtrl.Finish()
 	mockComputeClient := mocks.NewMockClient(mockCtrl)
@@ -133,7 +128,7 @@ func TestAPIInflater_Inflate_CreateDiskSuccess_CancelWithDeleteDisk(t *testing.T
 	mockLogger := mocks.NewMockLogger(mockCtrl)
 	mockLogger.EXPECT().Debug("apiInflater.inflate is canceled: cancel")
 
-	inflater := createAPIInflater(ImageImportRequest{
+	apiInflater := createAPIInflater(ImageImportRequest{
 		Source:       fileSource{gcsPath: "gs://bucket/vmdk"},
 		Subnet:       "projects/subnet/subnet",
 		Network:      "projects/network/network",
@@ -141,10 +136,7 @@ func TestAPIInflater_Inflate_CreateDiskSuccess_CancelWithDeleteDisk(t *testing.T
 		ExecutionID:  "1234",
 		NoExternalIP: false,
 		WorkflowDir:  daisyWorkflows,
-	}, mockComputeClient, &storage.Client{}, mockLogger)
-
-	apiInflater, ok := inflater.(*apiInflater)
-	assert.True(t, ok)
+	}, mockComputeClient, &storage.Client{}, mockLogger, true)
 
 	// Send a cancel signal in prior to guarantee cancellation logic can be executed.
 	cancelResult := apiInflater.Cancel("cancel")
@@ -154,7 +146,7 @@ func TestAPIInflater_Inflate_CreateDiskSuccess_CancelWithDeleteDisk(t *testing.T
 	assert.NoError(t, err)
 }
 
-func TestAPIInflater_Inflate_Cancel_CleanupFailedToVerify(t *testing.T) {
+func TestAPIInflater_ShadowInflate_Cancel_CleanupFailedToVerify(t *testing.T) {
 	mockCtrl := gomock.NewController(t)
 	defer mockCtrl.Finish()
 	mockComputeClient := mocks.NewMockClient(mockCtrl)
@@ -163,7 +155,7 @@ func TestAPIInflater_Inflate_Cancel_CleanupFailedToVerify(t *testing.T) {
 	mockLogger := mocks.NewMockLogger(mockCtrl)
 	mockLogger.EXPECT().Debug("apiInflater.inflate is canceled, cleanup failed to verify: cancel")
 
-	inflater := createAPIInflater(ImageImportRequest{
+	apiInflater := createAPIInflater(ImageImportRequest{
 		Source:       fileSource{gcsPath: "gs://bucket/vmdk"},
 		Subnet:       "projects/subnet/subnet",
 		Network:      "projects/network/network",
@@ -171,16 +163,13 @@ func TestAPIInflater_Inflate_Cancel_CleanupFailedToVerify(t *testing.T) {
 		ExecutionID:  "1234",
 		NoExternalIP: false,
 		WorkflowDir:  daisyWorkflows,
-	}, mockComputeClient, &storage.Client{}, mockLogger)
-
-	apiInflater, ok := inflater.(*apiInflater)
-	assert.True(t, ok)
+	}, mockComputeClient, &storage.Client{}, mockLogger, true)
 
 	cancelResult := apiInflater.Cancel("cancel")
 	assert.False(t, cancelResult)
 }
 
-func TestAPIInflater_Inflate_Cancel_CleanupFailed(t *testing.T) {
+func TestAPIInflater_ShadowInflate_Cancel_CleanupFailed(t *testing.T) {
 	mockCtrl := gomock.NewController(t)
 	defer mockCtrl.Finish()
 	mockComputeClient := mocks.NewMockClient(mockCtrl)
@@ -189,7 +178,7 @@ func TestAPIInflater_Inflate_Cancel_CleanupFailed(t *testing.T) {
 	mockLogger := mocks.NewMockLogger(mockCtrl)
 	mockLogger.EXPECT().Debug("apiInflater.inflate is canceled, cleanup is failed: cancel")
 
-	inflater := createAPIInflater(ImageImportRequest{
+	apiInflater := createAPIInflater(ImageImportRequest{
 		Source:       fileSource{gcsPath: "gs://bucket/vmdk"},
 		Subnet:       "projects/subnet/subnet",
 		Network:      "projects/network/network",
@@ -197,17 +186,14 @@ func TestAPIInflater_Inflate_Cancel_CleanupFailed(t *testing.T) {
 		ExecutionID:  "1234",
 		NoExternalIP: false,
 		WorkflowDir:  daisyWorkflows,
-	}, mockComputeClient, &storage.Client{}, mockLogger)
-
-	apiInflater, ok := inflater.(*apiInflater)
-	assert.True(t, ok)
+	}, mockComputeClient, &storage.Client{}, mockLogger, true)
 
 	cancelResult := apiInflater.Cancel("cancel")
 	assert.False(t, cancelResult)
 }
 
 func TestAPIInflater_getCalculateChecksumWorkflow(t *testing.T) {
-	inflater := createAPIInflater(ImageImportRequest{
+	apiInflater := createAPIInflater(ImageImportRequest{
 		Source:       fileSource{gcsPath: "gs://bucket/vmdk"},
 		Subnet:       "projects/subnet/subnet",
 		Network:      "projects/network/network",
@@ -215,13 +201,10 @@ func TestAPIInflater_getCalculateChecksumWorkflow(t *testing.T) {
 		ExecutionID:  "1234",
 		NoExternalIP: false,
 		WorkflowDir:  daisyWorkflows,
-	}, nil, &storage.Client{}, logging.NewToolLogger(t.Name()))
-
-	apiInflater, ok := inflater.(*apiInflater)
-	assert.True(t, ok)
+	}, nil, &storage.Client{}, logging.NewToolLogger(t.Name()), true)
 
 	w := apiInflater.getCalculateChecksumWorkflow("")
-	_, ok = w.Vars["compute_service_account"]
+	_, ok := w.Vars["compute_service_account"]
 	assert.False(t, ok)
 
 	apiInflater.request.ComputeServiceAccount = "email"
