+++ cli_tools/go.mod
@@ -7,11 +7,11 @@ require (
 	cloud.google.com/go/logging v1.2.0 // indirect
 	cloud.google.com/go/storage v1.13.0
 	cos.googlesource.com/cos/tools.git v0.0.0-20210104210903-4b3bc7d49b79 // indirect
-	github.com/GoogleCloudPlatform/compute-image-tools/daisy v0.0.0-20210204013302-8240807dedc1
+	github.com/GoogleCloudPlatform/compute-image-tools/daisy v0.0.0-20210204232849-5bf13ea3342e
 	github.com/GoogleCloudPlatform/compute-image-tools/mocks v0.0.0-20200414213327-359251a2c860
 	github.com/GoogleCloudPlatform/compute-image-tools/proto/go v0.0.0
 	github.com/GoogleCloudPlatform/osconfig v0.0.0-20210202205636-8f5a30e8969f
-	github.com/aws/aws-sdk-go v1.37.3
+	github.com/aws/aws-sdk-go v1.37.5
 	github.com/cenkalti/backoff/v4 v4.1.0
 	github.com/dustin/go-humanize v1.0.0
 	github.com/go-ole/go-ole v1.2.5
@@ -29,7 +29,7 @@ require (
 	github.com/minio/highwayhash v1.0.1
 	github.com/stretchr/testify v1.6.1
 	github.com/vmware/govmomi v0.24.0
-	go.chromium.org/luci v0.0.0-20210204170102-f25dc79418e4 // indirect
+	go.chromium.org/luci v0.0.0-20210204234011-34a994fe5aec // indirect
 	go.opencensus.io v0.22.6 // indirect
 	golang.org/x/crypto v0.0.0-20201221181555-eec23a3978ad // indirect
 	golang.org/x/oauth2 v0.0.0-20210201163806-010130855d6c // indirect
